test: cover missing pix webhook secret
